bug fixes for following and highscores
